import-service/requirements.txt
importlib-metadata==4.13.0
importlib-metadata<6
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
sqlalchemy
psycopg2-binary
redis
python-dotenv
requests
fastapi
uvicorn
dropbox


